+++ test/fork/OmnichainForkTestBase.sol
@@ -180,7 +180,7 @@ abstract contract OmnichainForkTestBase is TestBaseWorkflow {
 
         liqHelper = new OmnichainLiquidityHelper(poolManager);
 
-        suckerRegistry = new JBSuckerRegistry(jbDirectory(), jbPermissions(), multisig(), address(0));
+        suckerRegistry = new JBSuckerRegistry(jbDirectory(), jbPermissions(), jbPrices(), multisig(), address(0));
         hookStore = new JB721TiersHookStore();
         exampleHook = new JB721TiersHook(
             jbDirectory(),